embed navigation bar with tab bar
