+++ SmartBook/SmartBook/View/Authentication/LoginViewController.swift
@@ -29,15 +29,6 @@ class LoginViewController: UIViewController {
     }
     
     @IBAction func loginTapped(_ sender: UIButton) {
-        //        Router.show(
-        //            from: self,
-        //            storyboard: .userProfile,
-        //            identifier: "UserProfileViewController",
-        //            title: "User Profile",
-        //            embedInNavigation: true,
-        //            presentModally: true
-        //        )
-        
         let tabBar = MainTabBarController()
         tabBar.modalPresentationStyle = .fullScreen
         present(tabBar, animated: true)
