+++ SmartBook/SmartBook/View/CalendarViewController.swift
@@ -12,6 +12,6 @@ class CalendarViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        view.backgroundColor = .white
+        view.backgroundColor = .yellow
     }
 }
