+++ SmartBook/SmartBook/View/FavouriteViewController.swift
@@ -12,7 +12,7 @@ class FavouriteViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        view.backgroundColor = .white
+        view.backgroundColor = .blue
     }
 
 }
