+++ SmartBook/SmartBook/View/HomeViewController.swift
@@ -12,7 +12,7 @@ class HomeViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        view.backgroundColor = .white
+        view.backgroundColor = .green
     }
 
 }
