+++ SmartBook/SmartBook/View/SearchViewController.swift
@@ -12,7 +12,7 @@ class SearchViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        view.backgroundColor = .white
+        view.backgroundColor = .systemPink
     }
 
 }
